Factoring out test utility functions into package
